requirements.txt
django==3.0.7
pillow==7.1.2
arc4